backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
nltk==3.8.1
werkzeug==3.0.1